minor: use uint8_t type for integer width

+++ metagraph/src/cli/config/config.hpp
@@ -74,7 +74,8 @@ class Config {
     unsigned int port = 5555;
     unsigned int bloom_max_num_hash_functions = 10;
     unsigned int num_columns_cached = 10;
-    unsigned int count_width = 8;
+
+    uint8_t count_width = 8;
 
     // Alignment options
     bool alignment_seed_unimems = false;
